Move `TokenBalances` to separate module
